+++ packages/ripple-tide-api/src/nuxt.ts
@@ -10,6 +10,26 @@ import {
   installModule
 } from '@nuxt/kit'
 import rippleModules from './ripple-modules.js'
+import type { Options } from 'http-proxy-middleware'
+
+const extendNuxtProxyConfig = (nuxtConfig, proxyOption: Options) => {
+  let existingProxyItems: Options[] = []
+  if (nuxtConfig.options?.proxy?.options) {
+    if (Array.isArray(nuxtConfig.options?.proxy?.options)) {
+      existingProxyItems = [...nuxtConfig.options.proxy.options]
+    } else {
+      existingProxyItems = [nuxtConfig.options.proxy.options]
+    }
+  }
+
+  if (nuxtConfig.options.proxy) {
+    nuxtConfig.options.proxy.options = [...existingProxyItems, proxyOption]
+  } else {
+    nuxtConfig.options.proxy = {
+      options: [...existingProxyItems, proxyOption]
+    }
+  }
+}
 
 export default defineNuxtModule({
   meta: {
@@ -31,7 +51,10 @@ export default defineNuxtModule({
       content: {},
       site: ''
     },
-    debug: false
+    debug: false,
+    proxy: {
+      options: []
+    }
   },
 
   async setup(options: RplTideModuleConfig, nuxt) {
@@ -51,6 +74,19 @@ export default defineNuxtModule({
       site: await resolvePath('@dpc-sdp/ripple-tide-api/mapping/site')
     }
     nuxt.options.runtimeConfig.public.tide = options
+
+    const webformProxy = {
+      target: options.contentApi.baseUrl,
+      changeOrigin: true,
+      pathRewrite: {
+        '^/api/tide/': '/api/v1/'
+      },
+      pathFilter: ['/api/tide/webform_submission/**']
+    }
+
+    extendNuxtProxyConfig(nuxt, webformProxy)
+
+    await installModule('nuxt-proxy')
 
     // API endpoint handlers - See https://v3.nuxtjs.org/guide/directory-structure/server#api-routes
     addServerHandler({