+++ packages/ripple-tide-landing-page/src/mapping/page-components/webforms/webforms-mapping.ts
@@ -2,6 +2,9 @@ import { TideDynamicPageComponent } from '@dpc-sdp/ripple-tide-api'
 import { FormKitSchemaNode } from '@formkit/core'
 
 export interface ITideWebform {
+  formId: string
+  successMessageHTML: string
+  errorMessageHTML: string
   schema: FormKitSchemaNode
 }
 
@@ -15,9 +18,11 @@ export interface ITideFormElement {
 }
 
 const getFormSchemaFromMapping = async (
-  elements: ITideFormElement[],
+  webform,
   tidePageApi
 ): Promise<FormKitSchemaNode[]> => {
+  const elements: ITideFormElement[] = webform?.elements || []
+
   const getValidation = (
     field
   ): {
@@ -236,11 +241,12 @@ const getFormSchemaFromMapping = async (
       }
       case 'webform_actions':
         mappedField = {
-          $formkit: 'RplFormSubmit',
+          $formkit: 'RplFormActions',
           name: 'submit',
           variant: 'filled',
           label: field['#submit__label'],
           id: fieldKey,
+          displayResetButton: !!webform?.settings?.form_reset,
           ...getInputIcons(field)
         }
         break
@@ -266,13 +272,27 @@ export const webformMapping = async (
   tidePageApi
 ): TideDynamicPageComponent<ITideWebform> => {
   return {
-    component: 'RplForm',
+    component: 'TideLandingPageWebForm',
     id: field.drupal_internal__id,
     title: field.field_paragraph_title,
     props: {
-      id: field?.field_paragraph_webform?.drupal_internal__id,
+      formId: field?.field_paragraph_webform?.drupal_internal__id,
+      hideFormOnSubmit:
+        field?.field_paragraph_webform?.settings?.confirmation_type ===
+        'inline',
+      successMessageTitle:
+        field?.field_paragraph_webform?.settings?.confirmation_title ||
+        'Form submitted',
+      successMessageHTML:
+        field?.field_paragraph_webform?.settings?.confirmation_message ||
+        'Thank you! Your response has been submitted.',
+      errorMessageTitle: 'Form not submitted',
+      errorMessageHTML:
+        field?.field_paragraph_webform?.settings
+          ?.submission_exception_message ||
+        'We are experiencing a server error. Please try again, otherwise contact us.',
       schema: await getFormSchemaFromMapping(
-        field?.field_paragraph_webform?.elements,
+        field?.field_paragraph_webform,
         tidePageApi
       )
     }