+++ packages/ripple-ui-forms/src/plugin.ts
@@ -2,7 +2,7 @@ import {
   text,
   email,
   tel,
-  submit,
+  actions,
   checkbox,
   checkboxGroup,
   dropdown,
@@ -65,8 +65,8 @@ rplFormInputs.library = (node) => {
       return node.define(checkboxGroup)
     case 'RplFormDropdown':
       return node.define(dropdown)
-    case 'RplFormSubmit':
-      return node.define(submit)
+    case 'RplFormActions':
+      return node.define(actions)
   }
 }